fix: 리뷰반영 - file 분리 & view 구조 변경
