+++ iOS/Catalog/Catalog/SimilarQuotation/View/SimilarQuotationCard/SimilarHMGOption.swift
